Put out, target, and .idea folders in gitignore. Move weapons into weaponsabstractfactory package to start working those whole deal into an abstract factory. Fix typo in berserk method
