+++ .gitignore
@@ -1,1 +1,3 @@
-.idea/workspace.xml
+.idea/
+target/
+out/
