--- src/main/java/com/johnny/pack/age/controller/attack/Attackable.java
@@ -1,5 +1,0 @@
-package com.johnny.pack.age.controller.attack;
-
-public interface Attackable {
-    String attack();
-}
