--- src/main/java/com/johnny/pack/age/controller/attack/Berserkable.java
@@ -1,5 +1,0 @@
-package com.johnny.pack.age.controller.attack;
-
-public interface Berserkable {
-    String goBersek();
-}
