+++ src/main/java/com/johnny/pack/age/controller/attack/Fight.java
@@ -356,7 +356,7 @@ public class Fight {
     private void displayAttackDetails(Character attacker, Character victim,
                                       String severity, int damage){
         String form = switch(severity){
-            case "critical" -> attacker.getBerserkable().goBersek();
+            case "critical" -> attacker.getBerserkable().goBerserk();
             case "normal" -> attacker.getAttackable().attack();
             case "low" -> attacker.getScratchable().scratch();
             default -> form = "missed";
