+++ src/main/java/com/johnny/pack/age/controller/attack/baseattack/Attackable.java
@@ -1,0 +1,5 @@
+package com.johnny.pack.age.controller.attack.baseattack;
+
+public interface Attackable {
+    String attack();
+}
