+++ src/main/java/com/johnny/pack/age/controller/attack/heavyattack/Berserkable.java
@@ -1,0 +1,5 @@
+package com.johnny.pack.age.controller.attack.heavyattack;
+
+public interface Berserkable {
+    String goBerserk();
+}
