+++ src/main/java/com/johnny/pack/age/controller/attack/heavyattack/Hack.java
@@ -1,4 +1,4 @@
-package com.johnny.pack.age.controller.attack;
+package com.johnny.pack.age.controller.attack.heavyattack;
 
 public final class Hack implements Berserkable {
     private final String berserk;
@@ -14,7 +14,7 @@ public final class Hack implements Berserkable {
     }
 
     @Override
-    public String goBersek() {
+    public String goBerserk() {
         return this.berserk;
     }
 }
