+++ src/main/java/com/johnny/pack/age/controller/attack/scratch/Scratchable.java
@@ -1,4 +1,4 @@
-package com.johnny.pack.age.controller.attack;
+package com.johnny.pack.age.controller.attack.scratch;
 
 public interface Scratchable {
     String scratch();
