+++ src/main/java/com/johnny/pack/age/model/weaponabstractfactory/weapon/Equipable.java
@@ -1,4 +1,4 @@
-package com.johnny.pack.age.model.weapon;
+package com.johnny.pack.age.model.weaponabstractfactory.weapon;
 
 public interface Equipable {
     String weaponType();
