+++ src/test/java/com/johnny/pack/age/model/weapon/StingerTest.java
@@ -1,5 +1,6 @@
 package com.johnny.pack.age.model.weapon;
 
+import com.johnny.pack.age.model.weaponabstractfactory.weapon.Stinger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
